feature: made custom hook to toggle modal
