+++ src/components/CreateNewMessageBtn/CreateNewMessageBtn.module.css
@@ -1,4 +1,4 @@
-.compose-btn {
+.compose_button {
   color: #99a1b3;
   border: none;
   background: none;
