--- src/models/ModalFunction.ts
@@ -1,3 +1,0 @@
-export interface ModalFunction {
-  toggleModal: () => void;
-}
