+++ src/models/ModalFunctionality.ts
@@ -1,0 +1,5 @@
+export interface ModalFunctionality {
+  isOpen: boolean;
+  toggleModal: () => void;
+  handleSendMessage: () => void;
+}
